+++ libfuturize/fixer_util.py
@@ -12,8 +12,8 @@ from lib2to3.fixer_util import (FromImport, Newline, is_import,
                                 find_root, does_tree_import, Comma)
 from lib2to3.pytree import Leaf, Node
 from lib2to3.pygram import python_symbols as syms, python_grammar
-# from lib2to3.pgen2 import token
 from lib2to3.pygram import token
+import re
 
 
 ## These functions are from 3to2 by Joe Amenta:
@@ -195,7 +195,13 @@ def future_import(feature, node):
     if does_tree_import(u"__future__", feature, node):
         return
 
+    # Look for a shebang line
+    shebang_idx = None
+
     for idx, node in enumerate(root.children):
+        # If it's a shebang line, attach the prefix to
+        if is_shebang_comment(node):
+            shebang_idx = idx
         if node.type == syms.simple_stmt and \
            len(node.children) > 0 and node.children[0].type == token.STRING:
             # skip over docstring
@@ -209,6 +215,12 @@ def future_import(feature, node):
             return
 
     import_ = FromImport(u'__future__', [Leaf(token.NAME, feature, prefix=" ")])
+    if shebang_idx == 0 and idx == 0:
+        # If this __future__ import would go on the first line,
+        # detach the shebang prefix from the current first line
+        # and attach it to our new __future__ import node.
+        import_.prefix = root.children[0].prefix
+        root.children[0].prefix = u''
     children = [import_, Newline()]
     root.insert_child(idx, Node(syms.simple_stmt, children))
 
@@ -400,5 +412,15 @@ def check_future_import(node):
         # TODO: handle brackets like this:
         #     from __future__ import (absolute_import, division)
         assert False, "strange import: %s" % savenode
+
+
+SHEBANG_REGEX = ur'''^#!\s*.*python'''
+
+def is_shebang_comment(node):
+    """
+    Comments are prefixes for Leaf nodes. Returns whether the given node has a
+    prefix that looks like a shebang line.
+    """
+    return bool(re.match(SHEBANG_REGEX, node.prefix))